dashboard/requirements.txt
wheel
Jinja2
flask
Werkzeug
Pillow
